Clean up code, add packages + PKGBUILD and test hyprpm fix
The code has been cleaned up in a major way adding more steps instead of just using &&. Some packages have been added one requires a PKGBUILD as it has a dependency that doesn't need to be there. While the aur user exists we will run Hyprland in hopes that it solves the hyprpm issues.