Configs, future files and replace ironbar
Ironbar wasn't the best project but it had a few nice features and a large amount of people using it. We are now moving away from it and replacing it with ashell. Xava is being used as a cool way of seeing the audio from the desktop under the bar.  We also have big plans for themes later as we found better replacements for what we are currently doing.