Use EWW and add packages/configs
We are pulling a lot more packages all for different reasons but they will all be used. Removing nemo-pastebin and nemo-fileroller as we don't like pastebin and we are using a different fileroller. Swaync is being replace with hyprnotify and hyprshot is being replaced with satty. Xava had it's config reworked, eww is our new "bar".