Rotate cava, remove grap and install nwg-look
Cava with eww works great but it needs more bars, flipped vertically and remove the gap between bars. All of these have been solved with this commit and xava is no longer needed because of this. We are testing more with GTK themes so using nwg-look will be a great help to do that.