Packages, hyprland and themes
All packages we don't trust are going to be installed through flatpak. OBS also works better on flatpak. We have also installed a few other packages such as zsh (and packages that extend it's functionality) and lact. The custom theme has been renamed. Hyprland got it's default wallpaper set to number 0 and the news/donations have been disabled. We are testing walker as the app launcher. We hope we fixed ly and color_manager with this commit too.