Use hyprpolkit, fix icons and include matrix2051
We are installing the matrix2051 package so we can use an IRC client to connect to the matrix network. Pscircle is pointless since the commands don't function (too old I guess). The icons still had a few issues with the colours they picked and we hope this fixed it. The last thing was uploading the wallpapers which will be used with swww based on the clock at some point.