Basically a complete redo of everything (see description)
We have deleted the gtk theme stored in .themes since it never showed up as a valid gtk theme.

We moved all tmp files required for the building of the image into a tmp_build folder.

Almost all configs are now symlinks so that we can change them when we update the image.

The ly config is finally fixed which means a few things such as no empty password entrys, a clock and colour mixing animation.

We now are set the gtk theme with on config files (though not gtk 4.0 yet).

Bashrc includes the uri for libvirt so we don't need to run a script for it.

We can use the environment file to define global envs such as one which we have done which will skip clearing gl shaders for games (eg. the finals).

Added more colours for the colour-icons scripts, still not happy with the outcome just yet.

Provide systemd services, these will be used in future to run the first boot script and attach the network storage.

Dunst is included now as we plan to use it for notifications later but not just yet. We also include it's config

Eww has had it's cava changes revered since they made cava invisible. We now show seconds on the clock but it only updates every 5 seconds (also new icon). Correct font used now and the buttons are now square. Logout button now uses correct command to logout.

Hyprland config is split over many different files and has had changes such as stronger blur, using gnome-polkit now, more keybinds and other changes.

Wezterm is also fixed by making it fall back to x11 and it also uses the correct font now. It has a smaller zoom and tabs are hidden if there is only 1. Last change was that it will quit even if something in the terminal is running.

The Containerfile has gone through many changes such as removing duplicate lines and reducing the amount of steps we take.  This should make a smaller image since we use less steps and for aur packages remove make deps after done using them. We now clear up the packages which are not needed too. Many packages were added and removed (see diff)

This is a massive commit so I expect issues.