Remove the deletion of unused packages
Sadly it seems that we can't remove the unused package from inside of the image. I don't expect to be able to do it from a live system either because pacman won't work. I am letting it list all the packages though just so I can see if it's worth spending more time trying to clear them out.