Remove extra fonts, gamescope and fix osd and package cache clearing
We don't need the extra fonts as far as I know and they take up quite a bit of storage. Gamescope can be removed for now (maybe I'll add it back later) and we are going to use swayod rather than syshud. We don't need to build this package and it's a better more maintained package. When trying to clear the package cache we were only clearing the extra repos (this was none). We have fixed this by reverting a past change,