Fix hyprland config issues
The plugins config was never linked causing hyprland to be unable to find it. We are not going to use hyprscrolling because I love the master layout now. When using source in the hyprland config we should follow the example just to be safe. With the colours of the border they should be using rgb not rgba. Also we give the master window a bit more space. Last swww should run when hyprland opens also we fixed another spelling error