Move back to syshud and clear packages
We should be installing libadwaita-without-adwaita before we use anything that requires libadwaita. We are not going to use winapps any more but we still want to use a windows vm for some tasks, winapps just doesn't work well with libvirt and wayland. We are omving back to syshud and kitty instead of rio and swayosd (though I would like to keep swayosd if I can get it to look the same). We also want to get rid of the capturing region with screenshots.