Code cleanup and removal of millennium
Most of this commit is just clean up of lines that can be reduced to lower the amount of layers. Millennium seems broken from my limited testing both in the container and a separate desktop. We added the fastfetch line just to see what the output looks like and the amount of packages. Also I don't believe the assume-installed flag needs to be used. Cava needs a folder so we create that with a gitkeep file. There was a hyprland config file that shouldn't exist. We changed the colour of the ly colourmix (it's called colormix in the config) animation.