Rofi, oh-my-posh and cleanup
We are moving away from walker since it feels still too slow and we would rather use projects in the archlinux repos if we can. Rofi is a better project but requires more setup which has been done in this commit. Oh my posh is going to be used for our prompt but at some point we would like to rewrite the prompt for starship. We also changed the icon colours for the notifications to match the gold/gruvbox theme. Then to finish we just deleted more unused lines.