Reduce layers, revert part of ly, oh my posh, rofi and geforce now
We have made all the install-packages-build at the start just one long command to reduce layers. Ly has a colour fix and we reverted the centering of text. Rofi can now be opened with keybinds and is all in one config. It's also at the bottom and is short but wide. We also removed geforcenow electron since it isn't being worked on and we are planing to use the official package from nvidia