Move folder, back to bash and colour rofi
First we deleted the scripts folder since we don't really need it so everything is moved just to the tmp_build folder. We have removed zsh and don't need the packages for it, instead we have a bash config (inspired by chris titus tech). We added gamescope which might be used depending on how we find it works. Last we changed the colour of rofi (gruvbox is not blue)