Remove cava and some packages + add dnsmasq
Cava is just causing too many issues and doesn't function well. I would love to use xava but compiling it is a pain and doesn't seem to function even after compiling. We are going to get the linuxmint packages straight from the linuxmint repo so for now we have disabled those packages. dnsmasq  is an important package for virt-manager but at some point it was removed.